# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for retrieving time series data from stock markets",
        "title": "Stock Market API",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/timeseries/{symbol}/{interval}": {
            "get": {
                "description": "Returns intraday time series data with the specified interval",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timeseries"
                ],
                "summary": "Get intraday time series data with specific interval",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g., AAPL, MSFT)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "1min",
                            "5min",
                            "15min",
                            "30min",
                            "60min"
                        ],
                        "type": "string",
                        "description": "Time interval for data",
                        "name": "interval",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "compact",
                            "full"
                        ],
                        "type": "string",
                        "default": "compact",
                        "description": "Amount of data to return",
                        "name": "outputsize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json",
                            "csv"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "Data type for response",
                        "name": "datatype",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Whether to include extended hours data",
                        "name": "extended_hours",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Whether to adjust for split and dividend events",
                        "name": "adjusted",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Month for historical intraday data (YYYY-MM format)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/alphavantage.TimeSeriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/filereader/readbytes": {
            "post": {
                "description": "Reads and parses file data (CSV or PDF) from the request body and returns structured data",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filereader"
                ],
                "summary": "Read and parse file data from request body",
                "parameters": [
                    {
                        "enum": [
                            "csv",
                            "pdf"
                        ],
                        "type": "string",
                        "description": "File type (csv or pdf)",
                        "name": "fileType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Original filename (for reference only)",
                        "name": "filename",
                        "in": "query"
                    },
                    {
                        "description": "Raw file content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/filereader.FileReadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/filereader/upload": {
            "post": {
                "description": "Reads and parses an uploaded file (CSV or PDF) and returns structured data",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filereader"
                ],
                "summary": "Read and parse an uploaded file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to be parsed",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "csv",
                            "pdf"
                        ],
                        "type": "string",
                        "description": "File type (csv or pdf)",
                        "name": "fileType",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/filereader.FileReadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/timeseries/{symbol}": {
            "get": {
                "description": "Returns time series data for the specified stock symbol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timeseries"
                ],
                "summary": "Get time series data for a specific symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock symbol (e.g., AAPL, MSFT)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "TIME_SERIES_DAILY",
                            "TIME_SERIES_WEEKLY",
                            "TIME_SERIES_MONTHLY"
                        ],
                        "type": "string",
                        "default": "TIME_SERIES_DAILY",
                        "description": "Time series function",
                        "name": "function",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "compact",
                            "full"
                        ],
                        "type": "string",
                        "default": "compact",
                        "description": "Amount of data to return",
                        "name": "outputsize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json",
                            "csv"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "Data type for response",
                        "name": "datatype",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/alphavantage.TimeSeriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "alphavantage.TimeSeriesResponse": {
            "description": "Time series response data structure",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "interval": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "filereader.FileReadResponse": {
            "description": "File read response data structure",
            "type": "object",
            "properties": {
                "columnsMax": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "error": {
                    "type": "string"
                },
                "fileType": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "rowsCount": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    }
}